+++ CHANGES.rst
@@ -1,10 +1,13 @@
 Changelog
 =========
 
-1.4 (unreleased)
-----------------
+1.3.1 (unreleased)
+------------------
 
-- Nothing changed yet.
+- Added helper method 'utils.getPreviousEvent' that will receive an event
+  as parameter and will return the previous event in the workflow_history
+  if found
+  [gbastien]
 
 
 1.3 (2015-04-15)